refactor+security: extract components, optimize perf, fix security & bugs
- Extract I18N dictionary to app/lib/i18n.js (-365 lines from monolith)
- Extract FloatingLegalAgent + LEGAL_KB + ClotureMensuelle + PortailClientSS to modules
- Memoize dashboard KPI stats with useMemo (avoid 4 expensive reduce/filter on each render)
- Fix auto-backup useEffect missing s.activeClient dependency
- Fix undefined `emps` bug in EcheancierPaiements (use clients.flatMap)
- Fix wrong argument counts: generateBelcotaxXML, generateDimonaXML, generateDmfAXML, generateSEPAXML
- Fix generatePayslipPDF argument detection (auto-compute payroll when called with 2 args)
- Add escapeHtml utility for XSS protection in document.write() (printToPDF, attestations)
- Add auth to /api/push, /api/webhooks, /api/lois-update
- Add SSRF protection to /api/webhooks (validate webhook URLs, block private IPs)
- Remove unused detectAgentLangLazy import
- Add 79-test suite (tests/test-lib-extracts.js)
- Monolith: 24,762 → 22,947 lines (-1,815 lines)
