+++ app/lib/i18n.js
@@ -1,0 +1,368 @@
+// ── I18N — Dictionnaire FR / NL / EN / DE ──
+// Extrait de AureusSocialPro.js pour réduire la taille du monolithe
+export const I18N = {
+  // ── Navigation principale ──
+  'nav.dashboard': { fr:'Tableau de bord', nl:"Dashboard", en:"Dashboard", de:"Dashboard" },
+  'nav.employees': { fr:'Employés', nl:"Werknemers", en:"Employees", de:"Mitarbeiter" },
+  'nav.payslip': { fr:'Fiches de paie', nl:"Loonfiches", en:"Payslips", de:"Gehaltsabrechnungen" },
+  'nav.onss': { fr:'ONSS / Déclarations', nl:"RSZ / Aangiften", en:"Social Security", de:"Sozialversicherung" },
+  'nav.fiscal': { fr:'Fiscal', nl:"Fiscaal", en:"Tax", de:"Steuern" },
+  'nav.salaires': { fr:'Salaires & Calculs', nl:"Lonen & Berekeningen", en:"Salaries & Calculations", de:"Gehälter & Berechnungen" },
+  'nav.avantages': { fr:'Avantages & Rémunération', nl:"Voordelen & Verloning", en:"Benefits & Compensation", de:"Vorteile & Vergütung" },
+  'nav.contrats': { fr:'Contrats & Documents', nl:"Contracten & Documenten", en:"Contracts & Documents", de:"Verträge & Dokumente" },
+  'nav.rh': { fr:'RH & Personnel', nl:"HR & Personeel", en:"HR & Personnel", de:"HR & Personal" },
+  'nav.social': { fr:'Social & Assurances', nl:"Sociaal & Verzekeringen", en:"Social & Insurance", de:"Soziales & Versicherung" },
+  'nav.bienetre': { fr:'Bien-être & Prévention', nl:"Welzijn & Preventie", en:"Wellbeing & Prevention", de:"Wohlbefinden & Prävention" },
+  'nav.reporting': { fr:'Reporting & Export', nl:"Rapportage & Export", en:"Reporting & Export", de:"Berichterstattung & Export" },
+  'nav.legal': { fr:'Juridique & Veille', nl:"Juridisch & Monitoring", en:"Legal & Monitoring", de:"Recht & Überwachung" },
+  'nav.sprint9': { fr:'Sprint 9 - Modules', nl:'Sprint 9 - Modules', en:'Sprint 9 - Modules', de:'Sprint 9 - Module' },
+  'nav.portalmanager': { fr:'Portail Clients',nl:'Klantenportaal',en:'Client Portal',de:'Kundenportal' },
+    'nav.queue': { fr:'File traitement',nl:'Wachtrij',en:'Queue',de:'Warteschlange' },
+    'nav.onboarding': { fr:'Onboarding',nl:'Onboarding',en:'Onboarding',de:'Onboarding' },
+    'nav.cloture': { fr:'Clôture Mensuelle',nl:'Maandafsluiting',en:'Monthly Closing',de:'Monatsabschluss' },
+    'nav.notifications': { fr:'Notifications',nl:'Meldingen',en:'Notifications',de:'Benachrichtigungen' },
+    'nav.commandcenter': { fr:'Command Center',nl:'Command Center',en:'Command Center',de:'Command Center' },
+    'nav.autopilot': { fr:'Autopilot',nl:'Autopilot',en:'Autopilot',de:'Autopilot' },
+    'nav.smartauto': { fr:'Smart Auto',nl:'Smart Auto',en:'Smart Auto',de:'Smart Auto' },
+    'nav.massengine': { fr:'Mass Engine',nl:'Mass Engine',en:'Mass Engine',de:'Mass Engine' },
+    'nav.automatisation': { fr:'⚡ Automatisation',nl:'⚡ Automatisering',en:'⚡ Automation',de:'⚡ Automatisierung' },
+    'nav.team': { fr:'Équipe',nl:'Team',en:'Team',de:'Team' },
+  'nav.settings': { fr:'Paramètres', nl:"Instellingen", en:"Settings", de:"Einstellungen" },
+  'nav.aureussuite': { fr:'Aureus Suite', nl:"Aureus Suite", en:"Aureus Suite", de:"Aureus Suite" },
+  'nav.back': { fr:'← Tous les dossiers', nl:"← Alle dossiers", en:"← All folders", de:"← Alle Ordner" },
+  'nav.search': { fr:'🔍 Rechercher un module... (Ctrl+K)', nl:"🔍 Module zoeken... (Ctrl+K)", en:"🔍 Search module... (Ctrl+K)", de:"🔍 Modul suchen... (Ctrl+K)" },
+  'nav.noresult': { fr:'Aucun module trouvé', nl:"Geen module gevonden", en:"No module found", de:"Kein Modul gefunden" },
+  
+  // ── Sous-menus ──
+  'sub.dimona': { fr:'Dimona', nl:"Dimona" },
+  'sub.dmfa': { fr:'DMFA / DRS', nl:"DmfA / DRS" },
+  'sub.drs': { fr:'DRS / Documents C', nl:"DRS / C-documenten" },
+  'sub.onssapl': { fr:'ONSS-APL (DMFAPPL)', nl:"RSZ-PPL (DmfAPPL)" },
+  'sub.belcotax': { fr:'Belcotax 281.xx', nl:"Belcotax 281.xx" },
+  'sub.precompte': { fr:'Précompte 274', nl:"Bedrijfsvoorheffing 274" },
+  'sub.fiches_ext': { fr:'Fiches spéciales', nl:"Speciale fiches" },
+  'sub.co2': { fr:'Calcul CO2 véhicules', nl:"CO2-berekening voertuigen" },
+  'sub.atn': { fr:'🚗 ATN Véhicules', nl:"🚗 VAA Voertuigen" },
+  'sub.od': { fr:'O.D. Comptables', nl:"Boekhoudkundige OD" },
+  'sub.provisions': { fr:'Provisions', nl:"Voorzieningen" },
+  'sub.cumuls': { fr:'Cumuls annuels', nl:"Jaarlijkse cumulatie" },
+  'sub.netbrut': { fr:'Net → Brut', nl:"Netto → Bruto" },
+  'sub.simcout': { fr:'💰 Simulation coût', nl:"💰 Kostensimulatie" },
+  'sub.saisies': { fr:'Saisies-Cessions', nl:"Beslagen-Overdrachten" },
+  'sub.indexauto': { fr:'Index automatique', nl:"Automatische index" },
+  'sub.horsforfait': { fr:'Heures supplémentaires', nl:"Overuren" },
+  'sub.totalreward': { fr:'🏆 Total Reward', nl:"🏆 Total Reward" },
+  'sub.transport': { fr:'🚇 Transport domicile', nl:"🚇 Woon-werkverkeer" },
+  'sub.treizieme': { fr:'🎄 13ème mois', nl:"🎄 13de maand" },
+  'sub.css': { fr:'💎 Cotisation spéc. SS', nl:"💎 Bijzondere SS" },
+  'sub.bonusemploi': { fr:'🎯 Bonus emploi', nl:"🎯 Werkbonus" },
+  'sub.cheques': { fr:'Chèques-Repas', nl:"Maaltijdcheques" },
+  'sub.ecocmd': { fr:'Éco-chèques', nl:"Ecocheques" },
+  'sub.cafeteria': { fr:'Plan cafétéria', nl:"Cafetariaplan" },
+  'sub.cct90': { fr:'Bonus CCT 90', nl:"Bonus CAO 90" },
+  'sub.warrants': { fr:'Warrants', nl:"Warrants" },
+  'sub.budgetmob': { fr:'Budget mobilité', nl:"Mobiliteitsbudget" },
+  'sub.ecocircul': { fr:'Notes de frais', nl:"Onkostennota\'s" },
+  'sub.contrats2': { fr:'Contrats de travail', nl:"Arbeidsovereenkomsten" },
+  'sub.reglement': { fr:'Règlement de travail', nl:"Arbeidsreglement" },
+  'sub.compteindiv': { fr:'Compte individuel', nl:"Individuele rekening" },
+  'sub.preavis': { fr:'Préavis légal', nl:"Wettelijke opzegging" },
+  'sub.pecsortie': { fr:'Pécule de sortie', nl:"Vertrekvakantiegeld" },
+  'sub.certpme': { fr:'Certificat PME', nl:"KMO-certificaat" },
+  'sub.absences': { fr:'Gestion absences', nl:"Afwezigheidsbeheer" },
+  'sub.onboarding': { fr:'Onboarding', nl:"Onboarding" },
+  'sub.offboarding': { fr:'Offboarding', nl:"Offboarding" },
+  'sub.registre': { fr:'Registre du personnel', nl:"Personeelsregister" },
+  'sub.totalreward': { fr:'Total Reward', nl:"Total Reward" },
+  'sub.absenteisme': { fr:'📊 Analyse absentéisme', nl:"📊 Absenteïsme-analyse" },
+  'sub.credittemps': { fr:'Crédit-temps', nl:"Tijdskrediet" },
+  'sub.chomtemp': { fr:'⚠ Chômage temporaire', nl:"⚠ Tijdelijke werkloosheid" },
+  'sub.congeduc': { fr:'🎓 Congé-éducation payé', nl:"🎓 Betaald educatief verlof" },
+  'sub.rcc': { fr:'RCC / Prépension', nl:"SWT / Brugpensioen" },
+  'sub.outplacement': { fr:'Outplacement', nl:"Outplacement" },
+  'sub.pointage': { fr:'⏱ Pointage & Portail', nl:"⏱ Tijdsregistratie & Portaal" },
+  'sub.planform': { fr:'Plan de formation', nl:"Opleidingsplan" },
+  'sub.medtravail': { fr:'Médecine du travail', nl:"Arbeidsgeneeskunde" },
+  'sub.selfservice': { fr:'👤 Portail travailleur', nl:"👤 Werknemersportaal" },
+  'sub.assloi': { fr:'Assurance-Loi AT', nl:"Arbeidsongevallenverzekering" },
+  'sub.assgroupe': { fr:'Assurance Groupe', nl:"Groepsverzekering" },
+  'sub.syndicales': { fr:'Primes syndicales', nl:"Syndicale premies" },
+  'sub.allocfam': { fr:'Alloc. familiales', nl:"Kinderbijslag" },
+  'sub.caissevac': { fr:'Caisse de vacances', nl:"Vakantiekas" },
+  'sub.rentes': { fr:'Rentes', nl:"Renten" },
+  'sub.decava': { fr:'DECAVA', nl:"DECAVA" },
+  'sub.aidesemploi': { fr:"🎯 Aides à l'emploi", nl:"🎯 Tewerkstellingssteun" },
+  'sub.planglobal': { fr:'Plan global prévention', nl:"Globaal preventieplan" },
+  'sub.paa': { fr:'Plan action annuel', nl:"Jaarlijks actieplan" },
+  'sub.risquespsycho': { fr:'Risques psychosociaux', nl:"Psychosociale risico\'s" },
+  'sub.alcool': { fr:'Politique alcool/drogues', nl:"Alcohol-/drugsbeleid" },
+  'sub.elections': { fr:'🗳 Élections sociales', nl:"🗳 Sociale verkiezingen" },
+  'sub.organes': { fr:'CE / CPPT / DS', nl:"OR / CPBW / VA" },
+  'sub.accounting': { fr:'Accounting Output', nl:"Boekhoudkundige Output" },
+  'sub.bilanbnb': { fr:'Bilan Social BNB', nl:"Sociaal Verslag NBB" },
+  'sub.bilan': { fr:'Bilan Social', nl:"Sociaal Verslag" },
+  'sub.statsins': { fr:'Statistiques INS', nl:"INS Statistieken" },
+  'sub.sepa': { fr:'SEPA / Virements', nl:"SEPA / Overschrijvingen" },
+  'sub.peppol': { fr:'🔗 PEPPOL e-Invoicing', nl:"🔗 PEPPOL e-Facturatie" },
+  'sub.envoi': { fr:'Envoi documents', nl:"Documenten versturen" },
+  'sub.exportimport': { fr:'Export / Import', nl:"Export / Import" },
+  'sub.ged': { fr:'📁 GED / Archivage', nl:"📁 DMS / Archivering" },
+  'sub.alertes': { fr:'Alertes légales', nl:"Juridische waarschuwingen" },
+  'sub.secteurs': { fr:'Secteurs spécifiques', nl:"Specifieke sectoren" },
+  'sub.eta': { fr:'Relevés ETA', nl:"ETA-overzichten" },
+  'sub.docsjuridiques': { fr:'📄 Documents Juridiques', nl:"📄 Juridische Documenten" },
+  'sub.config': { fr:'Configuration société', nl:"Bedrijfsconfiguratie" },
+  'sub.fraisgestion': { fr:'💰 Frais de gestion', nl:"💰 Beheerskosten" },
+  
+  // ── Headers & Titles ──
+  'app.title': { fr:'AUREUS SOCIAL', nl:"AUREUS SOCIAL" },
+  'app.subtitle': { fr:'Logiciel de Paie Pro', nl:"Professionele Loonsoftware" },
+  'app.client': { fr:'Client', nl:"Klant" },
+  
+  // ── Login ──
+  'login.title': { fr:'Accès sécurisé', nl:"Beveiligde toegang" },
+  'login.create': { fr:'Créer un code PIN', nl:"PIN-code aanmaken" },
+  'login.enter': { fr:'Entrez votre code PIN', nl:"Voer uw PIN-code in" },
+  'login.btn': { fr:'Se connecter', nl:"Aanmelden" },
+  'login.create_btn': { fr:'Créer', nl:"Aanmaken" },
+  
+  // ── Employees ──
+  'emp.title': { fr:'Gestion du personnel', nl:"Personeelsbeheer" },
+  'emp.add': { fr:'Ajouter un travailleur', nl:"Werknemer toevoegen" },
+  'emp.firstname': { fr:'Prénom', nl:"Voornaam" },
+  'emp.lastname': { fr:'Nom', nl:"Naam" },
+  'emp.niss': { fr:'NISS', nl:"INSZ" },
+  'emp.birthdate': { fr:'Date de naissance', nl:"Geboortedatum" },
+  'emp.address': { fr:'Adresse', nl:"Adres" },
+  'emp.city': { fr:'Ville', nl:"Stad" },
+  'emp.zip': { fr:'Code postal', nl:"Postcode" },
+  'emp.startdate': { fr:"Date d'entrée", nl:"Datum indiensttreding" },
+  'emp.enddate': { fr:'Date de sortie', nl:"Datum uitdiensttreding" },
+  'emp.function': { fr:'Fonction', nl:"Functie" },
+  'emp.department': { fr:'Département', nl:"Afdeling" },
+  'emp.contract': { fr:'Type contrat', nl:"Contracttype" },
+  'emp.regime': { fr:'Régime', nl:"Regime" },
+  'emp.salary': { fr:'Salaire mensuel brut', nl:"Bruto maandloon" },
+  'emp.iban': { fr:'IBAN', nl:"IBAN" },
+  'emp.civil': { fr:'État civil', nl:"Burgerlijke staat" },
+  'emp.children': { fr:'Enfants à charge', nl:"Kinderen ten laste" },
+  'emp.status_employe': { fr:'Employé', nl:"Bediende" },
+  'emp.status_ouvrier': { fr:'Ouvrier', nl:"Arbeider" },
+  'emp.status_student': { fr:'Étudiant', nl:"Student" },
+  'emp.status_flexi': { fr:'Flexi-job', nl:"Flexi-job" },
+  'emp.sexe': { fr:'Sexe', nl:"Geslacht" },
+  'emp.sexe_m': { fr:'Masculin', nl:"Man" },
+  'emp.sexe_f': { fr:'Féminin', nl:"Vrouw" },
+  'emp.save': { fr:'Enregistrer', nl:"Opslaan" },
+  'emp.delete': { fr:'Supprimer', nl:"Verwijderen" },
+  'emp.active': { fr:'Actif', nl:"Actief" },
+  'emp.inactive': { fr:'Inactif', nl:"Inactief" },
+  
+  // ── Contract types ──
+  'ct.cdi': { fr:'CDI', nl:"COT" },
+  'ct.cdd': { fr:'CDD', nl:"CBD" },
+  'ct.interim': { fr:'Intérimaire', nl:"Uitzendkracht" },
+  'ct.student': { fr:'Étudiant (650h)', nl:"Student (650u)" },
+  'ct.replacement': { fr:'Remplacement', nl:"Vervanging" },
+  'ct.tpartiel': { fr:'Temps partiel', nl:"Deeltijds" },
+  'ct.full': { fr:'Temps plein', nl:"Voltijds" },
+  
+  // ── Civil status ──
+  'civil.single': { fr:'Isolé', nl:"Alleenstaand" },
+  'civil.married1': { fr:'Marié — 1 revenu', nl:"Gehuwd — 1 inkomen" },
+  'civil.married2': { fr:'Marié — 2 revenus', nl:"Gehuwd — 2 inkomens" },
+  'civil.cohabit': { fr:'Cohabitant', nl:"Samenwonend" },
+  'civil.divorced': { fr:'Divorcé', nl:"Gescheiden" },
+  'civil.widow': { fr:'Veuf/Veuve', nl:"Weduwe/Weduwnaar" },
+  
+  // ── Payslip ──
+  'pay.title': { fr:'Fiche de paie', nl:"Loonfiche" },
+  'pay.period': { fr:'Période', nl:"Periode" },
+  'pay.month': { fr:'Mois', nl:"Maand" },
+  'pay.year': { fr:'Année', nl:"Jaar" },
+  'pay.days': { fr:'Jours prestés', nl:"Gewerkte dagen" },
+  'pay.gross': { fr:'TOTAL BRUT', nl:"TOTAAL BRUTO" },
+  'pay.net': { fr:'NET À PAYER', nl:"NETTO TE BETALEN" },
+  'pay.onss': { fr:'Cotisations ONSS', nl:"RSZ-bijdragen" },
+  'pay.onss_worker': { fr:'ONSS travailleur', nl:"RSZ werknemer" },
+  'pay.onss_employer': { fr:'ONSS employeur', nl:"RSZ werkgever" },
+  'pay.pp': { fr:'Précompte professionnel', nl:"Bedrijfsvoorheffing" },
+  'pay.css': { fr:'Cotisation spéciale séc. soc.', nl:"Bijzondere bijdrage SZ" },
+  'pay.base': { fr:'Salaire de base', nl:"Basisloon" },
+  'pay.overtime': { fr:'Heures sup. (150%)', nl:"Overuren (150%)" },
+  'pay.sunday': { fr:'Heures dimanche (200%)', nl:"Zondaguren (200%)" },
+  'pay.night': { fr:'Heures nuit (125%)', nl:"Nachturen (125%)" },
+  'pay.bonus': { fr:'Prime', nl:"Premie" },
+  'pay.y13': { fr:'13ème mois', nl:"13de maand" },
+  'pay.sick': { fr:'Salaire garanti maladie', nl:"Gewaarborgd loon ziekte" },
+  'pay.advance': { fr:'Acompte', nl:"Voorschot" },
+  'pay.garnish': { fr:'Saisie', nl:"Beslag" },
+  'pay.transport': { fr:'Transport', nl:"Transport" },
+  'pay.expense': { fr:'Frais propres employeur', nl:"Eigen kosten werkgever" },
+  'pay.mv': { fr:'Chèques-repas', nl:"Maaltijdcheques" },
+  'pay.eco': { fr:'Éco-chèques', nl:"Ecocheques" },
+  'pay.employer_cost': { fr:'COÛT EMPLOYEUR', nl:"WERKGEVERSKOST" },
+  'pay.bonus_emploi': { fr:"Bonus à l'emploi", nl:"Werkbonus" },
+  'pay.bonus_fisc': { fr:'Bonus emploi fiscal', nl:"Fiscale werkbonus" },
+  'pay.red_struct': { fr:'Réduction structurelle', nl:"Structurele vermindering" },
+  'pay.print': { fr:'Imprimer', nl:"Afdrukken" },
+  'pay.pdf': { fr:'Télécharger PDF', nl:"PDF downloaden" },
+  'pay.calculate': { fr:'Calculer', nl:"Berekenen" },
+  
+  // ── Heures supplémentaires ──
+  'hs.overtime': { fr:'H. sup.', nl:"Overuren" },
+  'hs.sunday': { fr:'H. dimanche', nl:"Zondaguren" },
+  'hs.night': { fr:'H. nuit', nl:"Nachturen" },
+  'hs.fiscal': { fr:'H.sup fiscales (180h)', nl:"Overuren fiscaal (180u)" },
+  'hs.volont': { fr:'HS volont. brut=net (h)', nl:"Vrijw. overuren bruto=netto (u)" },
+  'hs.relance': { fr:'HS relance T1 (h)', nl:"Relance-overuren T1 (u)" },
+  
+  // ── Settings ──
+  'set.company': { fr:'Configuration société', nl:"Bedrijfsconfiguratie" },
+  'set.name': { fr:"Nom de l'entreprise", nl:"Bedrijfsnaam" },
+  'set.vat': { fr:'Numéro TVA', nl:"BTW-nummer" },
+  'set.bce': { fr:'Numéro BCE', nl:"KBO-nummer" },
+  'set.onss_num': { fr:'Numéro ONSS', nl:"RSZ-nummer" },
+  'set.nace': { fr:'Code NACE', nl:"NACE-code" },
+  'set.iban': { fr:'IBAN', nl:"IBAN" },
+  'set.bic': { fr:'BIC', nl:"BIC" },
+  'set.cp': { fr:'Commission paritaire', nl:"Paritair comité" },
+  'set.sec_soc': { fr:'Secrétariat social', nl:"Sociaal secretariaat" },
+  'set.address': { fr:'Adresse', nl:"Adres" },
+  
+  // ── ONSS / Dimona ──
+  'onss.dimona_in': { fr:'Dimona IN — Entrée en service', nl:"Dimona IN — Indiensttreding" },
+  'onss.dimona_out': { fr:'Dimona OUT — Sortie de service', nl:"Dimona OUT — Uitdiensttreding" },
+  'onss.worker_type': { fr:'Type de travailleur', nl:"Type werknemer" },
+  'onss.dmfa': { fr:'Déclaration DMFA', nl:"DmfA-aangifte" },
+  'onss.trimester': { fr:'Trimestre', nl:"Kwartaal" },
+  
+  // ── SEPA ──
+  'sepa.title': { fr:'SEPA / Virements bancaires', nl:"SEPA / Bankoverschrijvingen" },
+  'sepa.salaries': { fr:'Salaires', nl:"Lonen" },
+  'sepa.all': { fr:'TOUT (salaires + ONSS + PP)', nl:"ALLES (lonen + RSZ + BV)" },
+  'sepa.sal_only': { fr:'Salaires uniquement', nl:"Alleen lonen" },
+  'sepa.onss_only': { fr:'ONSS uniquement', nl:"Alleen RSZ" },
+  'sepa.pp_only': { fr:'PP + CSS uniquement', nl:"Alleen BV + BBSZ" },
+  'sepa.download': { fr:'💾 Télécharger .xml', nl:"💾 .xml downloaden" },
+  'sepa.email': { fr:'📧 Ouvrir dans Mail', nl:"📧 Openen in Mail" },
+  'sepa.copy': { fr:'📋 Copier le message', nl:"📋 Bericht kopiëren" },
+  
+  // ── Reporting ──
+  'rep.belcotax': { fr:'Belcotax 281.xx', nl:"Belcotax 281.xx" },
+  'rep.bilan': { fr:'Bilan Social', nl:"Sociaal Verslag" },
+  'rep.export': { fr:'Exporter', nl:"Exporteren" },
+  'rep.generate': { fr:'Générer', nl:"Genereren" },
+  
+  // ── Contrats ──
+  'ctr.type_cdi': { fr:'Contrat à durée indéterminée', nl:"Arbeidsovereenkomst voor onbepaalde duur" },
+  'ctr.type_cdd': { fr:'Contrat à durée déterminée', nl:"Arbeidsovereenkomst voor bepaalde duur" },
+  'ctr.type_student': { fr:"Convention d'occupation étudiant", nl:"Studentenovereenkomst" },
+  'ctr.type_flexi': { fr:'Contrat flexi-job', nl:"Flexi-job contract" },
+  'ctr.generate': { fr:'Générer le contrat', nl:"Contract genereren" },
+  'ctr.preview': { fr:'Aperçu', nl:"Voorbeeld" },
+  
+  // ── Veille légale ──
+  'legal.title': { fr:'Veille légale & Calendrier 2026', nl:"Juridische monitoring & Kalender 2026" },
+  'legal.filter': { fr:'Filtrer', nl:"Filteren" },
+  'legal.all': { fr:'Tout', nl:"Alles" },
+  'legal.deadlines': { fr:'📅 Échéances', nl:"📅 Deadlines" },
+  'legal.new2026': { fr:'⚖️ Nouveautés 2026', nl:"⚖️ Nieuw in 2026" },
+  'legal.upcoming': { fr:'🔮 Réformes à venir', nl:"🔮 Komende hervormingen" },
+  'legal.index': { fr:'📈 Indexations', nl:"📈 Indexeringen" },
+  'legal.urgent': { fr:'Urgentes', nl:"Dringend" },
+  'legal.plan': { fr:'À planifier', nl:"Te plannen" },
+  'legal.reminder': { fr:'Rappels', nl:"Herinneringen" },
+  'legal.watch': { fr:'Veille', nl:"Monitoring" },
+  'legal.institutions': { fr:'🏛 Institutions de référence', nl:"🏛 Referentie-instellingen" },
+  'legal.summary': { fr:'Résumé', nl:"Overzicht" },
+  'legal.cal_pp': { fr:'📅 Calendrier PP (FinProf)', nl:"📅 Kalender BV (FinProf)" },
+  'legal.cal_onss': { fr:'📅 Calendrier ONSS / DmfA', nl:"📅 Kalender RSZ / DmfA" },
+  
+  // ── Months ──
+  'months': {
+    fr:['Janvier',"Février","Mars","Avril","Mai","Juin","Juillet","Août","Septembre","Octobre","Novembre","Décembre"],
+    nl:['Januari',"Februari","Maart","April","Mei","Juni","Juli","Augustus","September","Oktober","November","December"],
+  },
+  
+  // ── Buttons & General ──
+  'btn.save': { fr:'Enregistrer', nl:"Opslaan" },
+  'btn.cancel': { fr:'Annuler', nl:"Annuleren" },
+  'btn.delete': { fr:'Supprimer', nl:"Verwijderen" },
+  'btn.edit': { fr:'Modifier', nl:"Bewerken" },
+  'btn.add': { fr:'Ajouter', nl:"Toevoegen" },
+  'btn.close': { fr:'Fermer', nl:"Sluiten" },
+  'btn.download': { fr:'Télécharger', nl:"Downloaden" },
+  'btn.print': { fr:'Imprimer', nl:"Afdrukken" },
+  'btn.generate': { fr:'Générer', nl:"Genereren" },
+  'btn.validate': { fr:'Valider', nl:"Bevestigen" },
+  'btn.search': { fr:'Rechercher', nl:"Zoeken" },
+  'btn.reset': { fr:'Réinitialiser', nl:"Resetten" },
+  'btn.copy': { fr:'Copier', nl:"Kopiëren" },
+  'btn.export': { fr:'Exporter', nl:"Exporteren" },
+  'btn.import': { fr:'Importer', nl:"Importeren" },
+  'btn.send': { fr:'Envoyer', nl:"Versturen" },
+  'btn.details': { fr:'Détail', nl:"Detail" },
+  'btn.yes': { fr:'Oui', nl:"Ja" },
+  'btn.no': { fr:'Non', nl:"Nee" },
+  
+  // ── Dashboard ──
+  'dash.welcome': { fr:'Bienvenue', nl:"Welkom" },
+  'dash.total_emp': { fr:'Travailleurs actifs', nl:"Actieve werknemers" },
+  'dash.mass_sal': { fr:'Masse salariale', nl:"Loonmassa" },
+  'dash.onss_due': { fr:'ONSS dû', nl:"RSZ verschuldigd" },
+  'dash.pp_due': { fr:'PP dû', nl:"BV verschuldigd" },
+  'dash.alerts': { fr:'Alertes', nl:"Waarschuwingen" },
+  'dash.next_deadline': { fr:'Prochaine échéance', nl:"Volgende deadline" },
+  
+  // ── Clients ──
+  'cli.title': { fr:'Gestion des dossiers clients', nl:"Beheer klantenbestanden" },
+  'cli.add': { fr:'Nouveau dossier', nl:"Nieuw dossier" },
+  'cli.name': { fr:'Nom du client', nl:"Klantnaam" },
+  'cli.open': { fr:'Ouvrir', nl:"Openen" },
+  'cli.emps': { fr:'travailleurs', nl:"werknemers" },
+
+  // ── English translations (Sprint 49) ──
+  'nav.back': { fr:'Retour aux clients', nl:'Terug naar klanten', en:'Back to clients' },
+  'nav.search': { fr:'Rechercher...', nl:'Zoeken...', en:'Search...' },
+  'nav.cloture': { fr:'Cloture mensuelle', nl:'Maandafsluiting', en:'Monthly closing' },
+  'nav.notifications': { fr:'Notifications', nl:'Meldingen', en:'Notifications' },
+  'nav.commandcenter': { fr:'Centre de commande', nl:'Commandocentrum', en:'Command center' },
+  'nav.autopilot': { fr:'Autopilot', nl:'Autopilot', en:'Autopilot' },
+  'nav.massengine': { fr:'Moteur de masse', nl:'Massa-engine', en:'Mass engine' },
+  'nav.queue': { fr:'File de traitement', nl:'Verwerkingswachtrij', en:'Processing queue' },
+  'nav.onboarding': { fr:'Onboarding', nl:'Onboarding', en:'Onboarding' },
+  'nav.portalmanager': { fr:'Gestion Portails', nl:'Portaalbeheer', en:'Portal management' },
+  'nav.team': { fr:'Equipe', nl:'Team', en:'Team' },
+  'emp.title': { fr:'Gestion des travailleurs', nl:'Werknemersbeheer', en:'Employee management' },
+  'emp.add': { fr:'Ajouter', nl:'Toevoegen', en:'Add' },
+  'emp.first': { fr:'Prenom', nl:'Voornaam', en:'First name' },
+  'emp.last': { fr:'Nom', nl:'Naam', en:'Last name' },
+  'emp.niss': { fr:'NISS', nl:'INSZ', en:'NISS' },
+  'emp.brut': { fr:'Brut mensuel', nl:'Bruto maandloon', en:'Gross monthly' },
+  'emp.net': { fr:'Net', nl:'Netto', en:'Net' },
+  'emp.status': { fr:'Statut', nl:'Statuut', en:'Status' },
+  'emp.employe': { fr:'Employé', nl:'Bediende', en:'Employee' },
+  'emp.ouvrier': { fr:'Ouvrier', nl:'Arbeider', en:'Worker' },
+  'pay.brut': { fr:'Salaire brut', nl:'Brutoloon', en:'Gross salary' },
+  'pay.onss': { fr:'ONSS personnelle', nl:'RSZ persoonlijk', en:'Social security' },
+  'pay.pp': { fr:'Précompte professionnel', nl:'Bedrijfsvoorheffing', en:'Withholding tax' },
+  'pay.net': { fr:'Salaire net', nl:'Nettoloon', en:'Net salary' },
+  'pay.cout': { fr:'Cout employeur', nl:'Werkgeverskost', en:'Employer cost' },
+  'pay.cheques': { fr:'Cheques-repas', nl:'Maaltijdcheques', en:'Meal vouchers' },
+  'pay.frais': { fr:'Frais propres', nl:'Eigen kosten', en:'Own expenses' },
+  'pay.bonus': { fr:'Bonus emploi', nl:'Werkbonus', en:'Employment bonus' },
+  'btn.save': { fr:'Sauvegarder', nl:'Opslaan', en:'Save' },
+  'btn.cancel': { fr:'Annuler', nl:'Annuleren', en:'Cancel' },
+  'btn.delete': { fr:'Supprimer', nl:'Verwijderen', en:'Delete' },
+  'btn.edit': { fr:'Modifier', nl:'Bewerken', en:'Edit' },
+  'btn.export': { fr:'Exporter', nl:'Exporteren', en:'Export' },
+  'btn.download': { fr:'Telecharger', nl:'Downloaden', en:'Download' },
+  'btn.print': { fr:'Imprimer', nl:'Afdrukken', en:'Print' },
+  'btn.generate': { fr:'Generer', nl:'Genereren', en:'Generate' },
+  'msg.saved': { fr:'Sauvegarde reussie', nl:'Succesvol opgeslagen', en:'Saved successfully' },
+  'msg.error': { fr:'Erreur', nl:'Fout', en:'Error' },
+  'msg.confirm': { fr:'Confirmer', nl:'Bevestigen', en:'Confirm' },
+  'msg.loading': { fr:'Chargement...', nl:'Laden...', en:'Loading...' },
+};
+
